tests/wpt/tests/css/css-fonts/parsing/font-optical-sizing-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Module Level 4: getComputedStyle().fontOpticalSizing</title>
<link rel="help" href="https://www.w3.org/TR/css-fonts-4/#propdef-font-optical-sizing">
<meta name="assert" content="font-optical-sizing computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value('font-optical-sizing', 'auto');
test_computed_value('font-optical-sizing', 'none');
</script>
</body>
</html>
